Border_Layout
Border Layout is used to divide the entire frame into 5 parts - north, south, east, west, center.
If one of the parts is not given, for example, if north is not defined then the other parts which are directly attached to it, i.e. east, west and center, will cover the north area also.