Key Listener
The Java KeyListener in the Abstract Window Toolkit (AWT) is a fundamental tool for achieving this. The KeyListener Interface is found in “java.awt.event” package. The KeyListener port in Java AWT is quite used to listen for keyboard events, such as key presses and key releases. 
The KeyListener port defines three methods that you must implement:
1) keyPressed(KeyEvent e)
2) keyReleased(KeyEvent e)
3) keyTyped(KeyEvent e)