Mouse Event Listener
MouseListener is an interface in java.awt.event package . Mouse events are of two types. MouseListener handles the events when the mouse is not in motion. There are five types of events that MouseListener can generate. There are five abstract functions that represent these five events. 
The abstract functions are : 
1) void mouseReleased(MouseEvent e) : Mouse key is released
2) void mouseClicked(MouseEvent e) : Mouse key is pressed/released
3) void mouseExited(MouseEvent e) : Mouse exited the component
4) void mouseEntered(MouseEvent e) : Mouse entered the component
5) void mousepressed(MouseEvent e) : Mouse key is pressed